% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pom_goal.R
\name{calculate_pom_goal}
\alias{calculate_pom_goal}
\title{calculate number of poms}
\usage{
calculate_pom_goal(travel_hrs = 0, community_hrs = 0)
}
\arguments{
\item{travel_hrs}{number of hours of teaching}

\item{community_hrs}{number of hours with other people
Returns goal number of poms the floor of $14 - t - 1.5 c$ where $t$ denotes the number of hours of travel, and $c$ the number of hours of _community_, where the lattter includes teaching, outreach, and research collaborations. That, is I aim to draw a distinction between self-directed time, and time where I have the benefit of external motivations.}
}
\description{
calculate number of poms
}
