% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-var.R
\name{d_var}
\alias{d_var}
\title{Variance of Cohen's d function}
\usage{
d_var(d, n1, n2)
}
\arguments{
\item{d}{Cohen's d.}

\item{n1}{First arm's n.}

\item{n2}{Second arm's n.}
}
\description{
Calculates the variance of Cohen's d (standardised mean difference).
}
