% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dontpanic-package.R
\docType{package}
\name{dontpanic}
\alias{dontpanic}
\alias{dontpanic-package}
\title{Coding equivalent of a towel.}
\description{
My helper functions are parked here. Use at own risk.
}
\keyword{package}
