% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpsterfire_progress.R
\name{dumpsterfire_progress}
\alias{dumpsterfire_progress}
\title{Visualise thesis chapter progress}
\usage{
dumpsterfire_progress()
}
\description{
Scrapes measures-chapters worksheet for latest update.
}
