% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mean.R
\name{get_mean}
\alias{get_mean}
\title{Generic mean function}
\usage{
get_mean(dist, par)
}
\arguments{
\item{dist}{A baseR distribution string (e.g., "lnorm").}

\item{par}{A list of parameters.}
}
\description{
Calculates the mean for an arbitrary distribution.
}
