% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Get data}
\usage{
get_measures(atracker_file)
}
\arguments{
\item{atracker_file}{An irritatingly manual process at the moment.}
}
\description{
This function gathers up my productivity data for `measured.` blog and `life's work`.
}
\details{
If it doesn't run, try [these instructions](https://github.com/jennybc/googlesheets/issues/343).
}
