% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample.R
\name{get_sample}
\alias{get_sample}
\title{Sample function}
\usage{
get_sample(n, dist, par)
}
\arguments{
\item{n}{Sample size.}

\item{dist}{A baseR distribution name, e.g. "lnorm".}

\item{par}{A list of parameters.}
}
\description{
Generate a sample from an arbitrary distribution
}
