% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sd.R
\name{get_sd}
\alias{get_sd}
\title{Generic Sd Function}
\usage{
get_sd(dist, par, this_mean)
}
\arguments{
\item{dist}{A baseR distribution string (e.g., "lnorm").}

\item{par}{A list of parameters.}
}
\description{
Calculates the standard deviation for an arbitrary distribution.
}
