% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Get Sample Summary Statistics}
\usage{
get_summary(n, dist, par)
}
\arguments{
\item{n}{Sample size.}

\item{dist}{A baseR distribution name, e.g. "lnorm".}

\item{par}{A list of parameters.}
}
\description{
Generates a random sample and returns a summary statistics tibble from that
sample.
}
