% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_read_all.R
\name{gs_read_all}
\alias{gs_read_all}
\title{Read all worksheets}
\usage{
gs_read_all(ss, delay = 4)
}
\description{
I was having issues with timeouts, but using this script seems to solve that.
}
\details{
from this issue: https://github.com/jennybc/googlesheets/issues/320#issuecomment-372076823
}
