% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lappy486_print.R
\name{lappy486_print}
\alias{lappy486_print}
\title{Print all files in a folder on lappy486}
\usage{
lappy486_print(path = "print/")
}
\description{
I haven't figured out how to print multiple files. When I try, it times out.
}
