% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeswork.R
\name{lifeswork}
\alias{lifeswork}
\alias{ask_gs}
\alias{get_daily_tasks}
\alias{get_workload_key}
\alias{get_signifiers}
\alias{wrangle}
\alias{wrangle.default}
\alias{wrangle.daily_tasks}
\title{lifeswork}
\usage{
ask_gs(for_ws)

get_daily_tasks()

get_workload_key()

get_signifiers()

wrangle(tasks)

\method{wrangle}{default}(tasks)

\method{wrangle}{daily_tasks}(tasks)
}
\arguments{
\item{sim}{A simulation object produced by \link{metasims}.}
}
\description{
These functions use \link{measures_gs_key} to access googlesheets, and produce my productivity tools.

Keeping this separate as I tweak the wrangling so often.
}
\details{
The idea is to access gs as little as possible. Multiple things are generated from each data set. So I think I'll need to do this as methods.
}
