% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look-a-look.R
\name{look_a_look}
\alias{look_a_look}
\title{LookALook function}
\usage{
look_a_look(df, n = 3)
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{n}{How many rows to display.}
}
\description{
Quick view of the head of a tibble.
}
