% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr_study.R
\docType{data}
\name{lotr_study}
\alias{lotr_study}
\title{Dummy set of studies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 383 rows and 2 columns.
}
\usage{
lotr_study
}
\description{
These data were obtained from
\href{https://www.kaggle.com/paultimothymooney/lord-of-the-rings-data/data}{kaggle}.
}
\keyword{datasets}
