% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr_study_hash.R
\name{lotr_study_hash}
\alias{lotr_study_hash}
\title{Hash studies with LOTR characters}
\usage{
lotr_study_hash(dat, study_col, sep = "_")
}
\arguments{
\item{dat}{A data frame with \code{study_col} column.}

\item{study_col}{Column that uniquely identifies the study.}

\item{sep}{Separator between name and year, defaults to "_".}
}
\description{
Replaces study identifier column with random combination of lord of the rings
character and a year, from \link{lotr_study} dataset found on kaggle.
}
\details{
Currently 373 names, so only coded for < 373 studies atm. In future, will
need to filter duplicates if want more.
}
