% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measured_data.R
\name{measured_data}
\alias{measured_data}
\title{Get data}
\usage{
measured_data(atracker_file)
}
\arguments{
\item{atracker_file}{An irritatingly manual process at the moment.}
}
\description{
This function gathers up my productivity data for `measured.` blog and `life's work`.
}
\details{
If it doesn't run, try [these instructions](https://github.com/jennybc/googlesheets/issues/343).
}
