% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measured_data_raw.R
\docType{data}
\name{measured_data_raw}
\alias{measured_data_raw}
\title{Measured data}
\format{A list of datasets.
\describe{
  \item{time_tracker_raw}{daily time records raw data from atracker}
  \item{time_tracker}{tidied column names and date objects}
  \item{measures}{google sheets records}
  \item{workload_key}{current pom goals}
  \item{operations}{phases of work}
  \item{daily}{manual daily records tracked in a googlesheet}
  }}
\usage{
measured_data_raw
}
\description{
Raw datasets scraped from googlesheets and imported from atracker iphone app.
}
\keyword{datasets}
