% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measured_wrangle.R
\name{measured_wrangle}
\alias{measured_wrangle}
\title{wrangle measured. data}
\usage{
measured_wrangle(mdat)
}
\arguments{
\item{df}{\code{\link{measured_data}} output}
}
\description{
wrangle measured. data
}
