% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_gs_key.R
\docType{data}
\name{measures_gs_key}
\alias{measures_gs_key}
\title{Googlesheets key for measures sheets}
\format{A googlesheets key.}
\usage{
measures_gs_key
}
\description{
Got this using \link[googlesheets:gs_url]{googlesheets::gs_url}. I must admit I copied and pasted the key.
}
\details{
Not sure which is better. Isn't url? But I had many too many requests problems before, I'm loathe to mess with it.
}
\keyword{datasets}
