% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_update.R
\name{measures_update}
\alias{measures_update}
\title{Update measures}
\usage{
measures_update()
}
\description{
Need a function to source the new data, rebuild the blog.
}
\details{
Assumes lifeswork is cloned into repo directory.
\enumerate{
\item Saves current working directory.
\item Sets current working directory to lifework.
\item Updates data.
\item Renders site.
\item Pushes to git.
\item Sets working directory back to current.
}
}
