% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_intensity.R
\name{mosaic_intensity}
\alias{mosaic_intensity}
\title{Mosaic plot of workload}
\usage{
mosaic_intensity(days = 1e+05, df = lifeswork:::wrangle_measures(),
  angle = 35, hjust = 1, ...)
}
\arguments{
\item{days}{Number of days' data to display.}

\item{df}{Dataframe of measures.' data.}
}
\description{
A mosaicplot of intensity achieved by category and day.
}
