% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_tracker.R
\name{ms_progress}
\alias{ms_progress}
\title{Progress converter}
\usage{
ms_progress(progress)
}
\arguments{
\item{progress}{A number from 1 to 8.}
}
\description{
Progress converter
}
\seealso{
Other manuscript trackers: 
\code{\link{ms_dat}()},
\code{\link{ms_report}()},
\code{\link{ms_three}()}
}
\concept{manuscript trackers}
