% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_tracker.R
\name{ms_report}
\alias{ms_report}
\title{Manuscript summary}
\usage{
ms_report(ms, ms_url = Sys.getenv("MS_TRACKER"))
}
\arguments{
\item{ms}{manuscript and name of sheet}

\item{ms_url}{googlesheet of manuscript trackers}
}
\description{
Manuscript summary
}
\seealso{
Other manuscript trackers: 
\code{\link{ms_dat}()},
\code{\link{ms_three}()}
}
\concept{manuscript trackers}
