% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_tracker.R
\name{ms_vis}
\alias{ms_vis}
\title{Visualise manuscript progess over time}
\usage{
ms_vis(all_ms = FALSE, ms_url = Sys.getenv("MS_TRACKER"))
}
\arguments{
\item{all_ms}{Show active manuscripts, default, or all manuscripts.}

\item{ms_url}{googlesheet of manuscript trackers}
}
\description{
Visualise manuscript progess over time
}
