% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_mine.R
\name{msg}
\alias{msg}
\title{Charles' custom messages}
\usage{
msg(msg)
}
\arguments{
\item{msg}{Message.}
}
\value{

}
\description{
Don't forget to preface with dontpanic:: so can keep track of where it comes from.
}
\examples{
msg("quack quack")
}
