% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda-plot.R
\name{panda}
\alias{panda}
\title{Make a panda}
\usage{
panda(msg = NULL, seed = "random")
}
\arguments{
\item{msg}{The panda can speak your words.}

\item{seed}{Set a random seed to reproduce a panda.}
}
\description{
Make a panda
}
