% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poms.R
\name{pom_report}
\alias{pom_report}
\title{Retrieve poms for a given period}
\usage{
pom_report(time_period = 1, terse = TRUE, with_date = FALSE)
}
\description{
Retrieve poms for a given period
}
\concept{lifeswork}
