% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomeclature.R
\docType{data}
\name{quaver}
\alias{quaver}
\title{musical notation shortcuts for output in .Rmd; use as in-line R.}
\format{An object of class \code{knit_asis} of length 1.}
\usage{
quaver
}
\description{
musical notation shortcuts for output in .Rmd; use as in-line R.
}
\keyword{datasets}
