% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-within.R
\name{s_within}
\alias{s_within}
\title{S_within function}
\usage{
s_within(s1, s2, n1, n2)
}
\arguments{
\item{s1}{First arm's sd.}

\item{s2}{Other arm's sd.}

\item{n1}{First arm's n.}

\item{n2}{Second arm's n.}
}
\description{
Calculates the s_within for Cohen's d.
}
