% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_summary.R
\name{sample_summary}
\alias{sample_summary}
\title{Get Sample Summary Statistics}
\usage{
sample_summary(n, dist, par)
}
\arguments{
\item{n}{Sample size.}

\item{dist}{A baseR distribution name, e.g. "lnorm".}

\item{par}{A list of parameters.}
}
\description{
Returns a random sample with that sample's summary statistics presented.
}
