% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\title{Visualise between-study variability from a multinma model}
\usage{
tau(nma_mod)
}
\arguments{
\item{nma_mod}{A multinma model object.}
}
\description{
Visualise between-study variability from a multinma model
}
