% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetracker_path.R
\docType{data}
\name{timetracker_path}
\alias{timetracker_path}
\title{timetracker data}
\format{An object of class \code{character} of length 1.}
\usage{
timetracker_path
}
\description{
timetracker data
}
\keyword{datasets}
