% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_date.R
\name{title_date}
\alias{title_date}
\title{Format Date for .RPres}
\usage{
title_date(this_date = NULL)
}
\arguments{
\item{this_date}{Optional argument to set date with a `date` object.
Otherwise defaults to current date.}
}
\description{
Return a nicely formatted title date for RMarkdown presentations.
}
\examples{
# title_date()
# title_date("2017-11-04")
}
