% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today_date.R
\name{today_date}
\alias{today_date}
\title{Return date in language I'm trying to learn right now}
\usage{
today_date(language = "french", verbose = FALSE)
}
\description{
Current is French, but I might switch back to German. Depends on how the write-up is going.
}
\details{
Haven't coded this for other locations yet, but can be extended later.
}
