% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today_rethinking_ex.R
\name{today_rethinking_ex}
\alias{today_rethinking_ex}
\title{Randomly assign an exercise from statistical rethinking}
\usage{
today_rethinking_ex(e, m, h)
}
\arguments{
\item{e}{Vector of exercises, i.e., c(1, 2, 4, 5)}
}
\description{
Randomly assign an exercise from statistical rethinking
}
