% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_measures.R
\name{wrangle_measures}
\alias{wrangle_measures}
\title{wrangle measured. data}
\usage{
wrangle_measures(timetracker)
}
\arguments{
\item{df}{\code{\link{measured_data}} output}
}
\description{
this is currently hardcoded in document
}
