% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_dist.R
\name{any_dist}
\alias{any_dist}
\title{Quantile, denisty, or probatility for an arbitrary distribution.}
\usage{
any_dist(x, dist, par = NULL, type = "q")
}
\arguments{
\item{x}{A numeric quantile or probability.}

\item{dist}{A distribution in baseR speak, e.g. "lnorm".}

\item{par}{A `list` comprising parameters for the distribution. If
unspecified, default parameters will be used.}

\item{type}{"q" for quantile, "d" for density, "p" for probability.
Defaults to "q".}
}
\description{
Returns a quantile, probability, or density for given distribution.
}
