% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Get summary statistics of a given sample}
\usage{
get_summary(sample)
}
\arguments{
\item{sample}{A numeric vector.}
}
\description{
Summary statistics of a given sample vector returned as a tibble.
}
