% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_tracker.R
\name{ms_dat}
\alias{ms_dat}
\title{#'
#'
#'
#'}
\usage{
ms_dat(ms, ms_url = Sys.getenv("MS_TRACKER"))
}
\arguments{
\item{ms}{manuscript and name of sheet}

\item{ms_url}{googlesheet of manuscript trackers}
}
\description{
gs_url <- "https://docs.google.com/spreadsheets/d/1sWRaWCw-grpf3AFktsRahFb03z37J0JE7mZxDp34FxM/edit#gid=1892813229"
}
\details{
gs_metadat <-
gs4_get()

gs_metadat \%>\%
pluck("sheets")

dat \%>\% filter(progress_achieved == min(progress_achieved)) \%>\% pluck("ms") \%>\% sample(3)
Get manuscript tracker data
}
\seealso{
Other manuscript trackers: 
\code{\link{ms_progress}()},
\code{\link{ms_report}()},
\code{\link{ms_three}()}
}
\concept{manuscript trackers}
