% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda-plot.R
\name{panda}
\alias{panda}
\title{Make a panda}
\usage{
panda(seed = "random", msg = NULL)
}
\arguments{
\item{seed}{Set a random seed to reproduce a panda.}

\item{msg}{The panda can speak your words.}
}
\description{
Make a panda
}
