% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature.R
\name{quaver}
\alias{quaver}
\title{musical notation shortcuts for output in .Rmd; use as in-line R.}
\usage{
quaver()
}
\description{
See \href{https://lifes-work.netlify.com/rituals.html#signifiers}{signifiers}.
}
